Added CHB LL & FFT stanzas to the train_mac.json and test_mac.json.
Tested the LSTM, DNN models with training on the CHB LL and CHB FFT datasets.
Tested the LSTM, DNN, and HYBRID models with testing on the
CHB LL and CHB FFT datasets.
This testing only proves that data engineering and model creation
part are working.  It does not currently produce good results and does
not prove that the current specs of the LSTM and DNN are good.